fix issues on jdk9+
